# Improve Axis3 interactivity performance