# Description

### Change in text

I noticed that `Axis3` can be rather sluggish at times and one reason for it seem to be text updates. A large part of the time spent on updating the angles in `Axis3` was in the update of `camera.projectionview`. From there I intercepted the updates with timed updates with this:

```julia
using GLMakie, LaTeXStrings
fig = Figure()
ax = Axis3(fig[1, 1], xlabel = L"Test $\frac{1}{1+x}$")

function print_time(f, i)
    x -> begin
        t = time()
        val = f(x)
        dt = time() - t
        if dt > 1e-3
            printstyled("\nTime: $dt\n", color = :red)
            println("@ index $i")
            flush(stdout)
        end
        return val
    end
end

src = ax.scene.camera.projectionview
for i in eachindex(src.listeners)
    src.listeners[i] = print_time(src.listeners[i], i)
end

fig
```

and found that most of the time was spend in `add_ticks_and_ticklabels!`, in (and after) `label_pos_rot_valign`. (You can find that out by doing autocompletes on the functions held by the observable.)

Well, long story shot - I removed the `position` trigger from text layouting and that reduced this particular timing from `~0.06s` to `~0.006s`.  I also changed the method with `(string, pos)` pairs to only update when necessary to avoid rebuilding glyphcollections too much.

### In Axis3

For this I tested with `src = ax.azimuth`. After the text changes the first 3 listeners took ~0.006, ~0.006 and ~0.012s. The first two are `mi1` and `mi2` and the last is the calculation of `projectionview`. I changed `mi1`, `mi2` and `mi3` to only update on changes which practically eliminates the first two calls.

I also split up some observables for the axis label to avoid triggering text layouting unnecessarily. That (or the changes `text!` for latexstrings) reduced the last timing to ~0.006s for me.


## Type of change

- Performance