@ffreyer 2022-05-04
It would probably be worth testing updates to attributes like `xticklabelalign` as well. I think those are the most likely to stop working from this