@ffreyer 2022-05-09
https://github.com/JuliaPlots/Makie.jl/blob/b3b0c6daaff5a7c5d5ed36310e5a65fc11832b67/GLMakie/src/drawing_primitives.jl#L313-L327

This is what makes updates of textsize, align, font and probably everything else that isn't position, offset or transfunc fail with `Vector{GlyphCollection}`. Why is it there?