@SimonDanisch 2022-05-11
>Somehow this changed the colors in "Dynamically adjusting number of particles in a meshscatter".

ColorTypes.jl changed their rand implementation, which lead to this... I rebased against master and updated the ref images, so hopefully should turn green now!